jobtap: don't set errno on success in jobtap_lookup_jobid()
Problem: For an unknown reason, if FLUX_JOBTAP_CURRENT_JOB is passed to
jobtap_lookup_jobid(), or the id argument matches the current jobid,
errno is set to EINVAL before returning the current job. Possibly
this is just a cut-and-paste error from the original implementation
of the function, since it doesn't make sense.

Do not set errno in this case of successful return from
jobtap_lookup_jobid().