dependency-after: support `afterexcept` dependency scheme
Problem: The dependency-after plugin supports `afternotok`, which runs
a job after the antecedent fails for any reason. However, in many cases
it may be more practical to only run a job after a job fails with a
fatal job exception, such as a node failure or timeout condition.

Add support for the `afterexcept` dependency scheme, which is only
satisfied when the antecedent jobs fails with a fatal exception.