libflux: simplify getting default URI in `flux_open_ex(3)`
Problem: The code to obtain the default URI when no URI is specified
is embedded in `flux_open_ex(3)` and can't be reused.

Move this code to a new internal function. As a side effect, the value
of the default URI is always copied, so rename the local variable
containing the allocated memory.