jobtap/killbot: add new preemption plugin
Problem: RFC 14 specifies that a job can set the preemptible-after
system attribute in their jobspec to signal that the job can be
preempted, but flux-core and flux-sched currently ignore this.

The scheduler is the proper place to handle preemption, but in the
mean time, provide a jobtap plugin that can do a sloppy job of it
without knowledge of the schedule.