libflux: accept ancestor path as URI alternative in `flux_open(3)`
Problem: It is awkward for users using the C or Python API to open
a Flux handle to an ancestor instance from within a child job. The
caller must first open a local handle, then fetch the `parent-uri`
attribute, close the local handle, and finally open a new handle to
the parent. These steps have to be repeated for traversing further
up the hierarchy. A simplified approach is needed to improve the
user experience.

Simplify access to ancestor Flux instances by enabling relative paths
in `flux_open(3)`. Use "/" for the root instance, ".." for the parent,
and "../.." for its parent, etc.  A single "." represents the current
instance and is ignored in path elements. For consistency with the
filesystem idiom, ".." at the top level is not an error and connects
to the root instance.

Examples:
 - `flux_open ("/")` returns a handle to the root instance
 - `flux_open (".")` returns a handle to the current instance, it is
   equivalent to `flux_open (NULL)`.
 - `flux_open ("..")` returns a handle to the parent
 - `flux_open ("../..")` returns a handle to the parent