kvs: report error to all fence requests
Problem: When an error occurs in a fence request, the error is
only reported to the requestor.  In most cases, this is fine, but
the error should be reported to all members of the fence in some
cases.  Most notably, it should be reported to all members of the
fence when the fence count has been reached and an error has
occurred after this point.

Solution: In fence_request_cb() return an error via
the kvs function error_event_send_to_name() when it is appropriate.