job-exec: fix confusing "job shell exec error" log message
Problem: When the job-exec module detects an exec error for a job
shell it emits a confusing error message that includes either
the path to the job shell or the IMP (if a multiuser job), and
only the result of `strerror()` for the errno returned from
libsubprocess. When using sdexec, this errno is always `ENOENT`,
resulting in a confusing error message that seems to indicate
that `flux-imp` was not found.

It is unhelpful to include `argv[0]` in this error message. It will
always be the job shell or the IMP and we all know it. Drop this
from the log message.

Also, sdexec will provide extra information in the subprocess error
string available from `flux_subprocess_fail_error (p)`. Log this
instead of `strerror (errno)`.

Fixes #6568