t/kvs: do not immediately exit on fence error
Problem: In the fence_api testing tool, if an error occurs during
a fence, the tool immediately exits.  This makes it difficult to
test for fence error conditions.

Solution: Do not exit immediately on error, capture the error and
call pthread_exit().  Determine if an error has occurred after
all threads have ended.