kvs: support fence-max-ops
Problem: The new transaction-max-ops configuration limits the number
of operations a user can submit in a KVS commit or fence.  However,
this does not cap the total number of operations from a combined
fence, which could exceed the transaction-max-ops by a lot.

Support a new KVS configuration "fence-max-ops" that will reject
KVS fences with a combined number of options above a maximum.  The
default maximum is 1048576.