resource: keep resource eventlog events in memory
Problem: the full resource eventlog, including online/offline
events that are not committed to the KVS, may need to be monitored.

Keep events in a json array in memory, including the events that
were read from the KVS at startup, if any.

Filter out any historical resource-define events.  These are meant for
synchronization on the availability of R and that only pertains to the
current instance.