shell: rcalc: sort ranks if unsorted in Rv1 R_lite array
Problem: The shell rcalc code assigns shell ranks by default in the
order they appear in the R_lite array in Rv1, but when these ranks
appear out of ascending order, this breaks assumptions elsewhere in
the shell that shell ranks are assigned in the same order as broker
ranks and the R nodelist.

Since the common case will be a sorted R_lite array, detect if the
ranks are not sorted and, if so, sort the rcalc rank array by broker
rank and reassign shell ranks.

Fixes #6582