resource: give reslog access to module context
Problem: the reslog class has no way to access the resource inventory,
but it will be useful to send journal consumers a copy of R when
the resource-define event is emitted.

Pass the resource_ctx to reslog_create() instead of just the flux_t
handle.  Adjust internal uses of the flux_t handle to get it via
reslog->ctx->h instead of reslog->h.