kvs: remove fence support
Problem: The KVS fence implementation is virtually unused.  It also
has the potential for denial-of-service attacks.  Its implementation
is also quite "one off" within the KVS, limiting some ability to refactor
the code.  For all these reasons, it should be removed.  If fence-like
behavior is ever needed again in the future, it should be implemented
as a stand alone module/service on top of the KVS.

Solution: Remove all fence support in the KVS.

Fixes #6587