cmd: only prepend to PYTHONPATH in `flux python` and `flux env`
Problem: Users that run Python programs using a different version or
installation of Python under `flux run` currently have to work around
the fact that `flux(1)` always prepends the path to the current Python
bindings to `PYTHONPATH`. This gets exported in the job environment
and the incorrect Flux Python bindings are loaded, causing errors.

The `flux(1)` command ostensibly prepends to `PYTHONPATH` to ensure
that Python subcommands use the right version of the bindings, but
the `py-runner.py` wrapper now takes care of that, so doing this is
actually no longer necessary.

Do not modify `PYTHONPATH` in `flux` unless `FLUX_PYTHONPATH_PREPEND`
is set. For backwards compatibility, continue to prepend the builtin
Python path in the `flux python` and `flux env` utilities.

Fixes #6594

fixup! cmd: do not always prepend to `PYTHONPATH` in `flux(1)`