cmd: save and restore original PYTHONPATH for python subcommands
Problem: The py-runner.py script is invoked by flux(1) for any Python
subcommand in order to ensure the subcommand has a sane sys.path
matching the currently invoked version of Flux. Ideally, this would
not require flux(1) to modify PYTHONPATH, which is then inherited by
the user's environment (e.g. in the case of `flux run`).  However,
`py-runner.py` does need the correct PYTHONPATH at startup so it can
`import flux`, so PYTHONPATH must be set.

In order to avoid polluting the user's environment, save the original
callers PYTHONPATH in FLUX_PYTHONPATH_ORIG in flux(1), then restore
it if set in `py-runner.py`. If FLUX_PYTHONPATH_ORIG is not set, this
means that the callers PYTHONPATH was empty, so unset the variable
in this case.