sdbus: add system module option
Problem: the sdbus module is hardwired to connect to a systemd
user instance, but Flux now has "work" running in the systemd
system instance as well (prolog, epilog, housekeeping).

Add a "system" module option which directs sdbus to connect
to the systemd system instance instead.  Future commits will allow
a second instance of the sdbus module to be loaded with this option
so access to both systemd instances can be handled concurrently.