sdmon: add new systemd monitor module
Problem: there is no mechanism to track systemd units across
a broker restart.

Add a broker module that creates and maintains a list of running
flux systemd units.

This monitors two instances of systemd:
- the user one, running as user flux (where jobs are run)
- the system one (where housekeeping, prolog, epilog run)

A list of units matching flux unit globs is requested at initialization,
and a subscription to property updates on those globs is obtained.
After the initial list, monitoring is driven solely by property updates.

Join the sdmon.online broker group once the node is demonstrably idle.
This lets the resource module on rank 0 notice compute nodes that need
cleanup at restart and withhold them from the scheduler.  Once the group
is joined, sdmon does not explicitly leave it.  It implicitly leaves the
group if sdmon is unloaded or the node goes offline/lost.

If there are running units at startup, log this information at LOG_ERR
level, and again when the units are cleaned up, e.g.

 flux-housekeeping@fAMpzjUPvsR.service needs cleanup - resources are offline
 cleanup complete - resources are online

In the future, this module's role could be expanded to support tooling
for listing running work and obtaining runtime information such as
pids and cgroup resource parameters.  It could also play a role in
informing other flux components about work that should be re-attached
after a full or partial restart, when support for that is added.