kvs: remove treq_t convenience type
Problem: The treq_t convenience type for storing a transaction request
is now a struct of just two values: a message request and a name created
for that request.  Having a unique struct and an API for it is now excessive.

Solution: Remove the treq_t type and API for its creation and
access of internal variables.  Remove all related unit tests.

Change the treq_mgr_add_transaction() function take the message request and
a name.  Have treq_mgr_lookup_transaction() return a flux_msg_t now.

Update tests and users accordingly.