doc: reword initial program environment documentation
Problem: The section of flux-environment(7) detailing the initial
program environment can be confusing to users because the set of
environment variables *not* set is prominently listed first in a
nice bulleted list, while environment variable(s) that *are*
set appear later in paragraph form. This may lead users that are
casually scanning the document to assume that the listed environment
variables will be found in the initial program environment.

Swap the section so that the environment variables that are set and
the suggested commands to obtain common values using Flux commands are
listed first and add documentation to emphasize their importance.

De-emphasize the unset variables by listing them in paragraph form
instead of a bulleted list.