kvs: remove stats clearing
Problem: KVS stats clearing code was added very early on in KVS
prototyping for testing.  It is basically not used anymore.  If we
want to test something that requires a "clean" slate for the KVS,
we can just start a new instance and test with that.

Solution: Remove all KVS stats clearing code.  Remove all associated
tests.

Fixes #6607