kvs: remove dropcache
Problem: KVS dropcache code was added very early on in KVS
prototyping for testing.  It is basically not used anymore.  If we
want to test something that requires a "clean" slate for the KVS,
we can just start a new instance and test with that.

Solution: Remove all KVS dropcache code.  Remove all associated
tests.  Update one regression test to be accomplished through two
instances rather than one to emulate an emptying of the cache.

Fixes #6608