kvs: do not pass lookup handle to getroot()
Problem: A lookup handle is passed to getroot() only because it
has to be passed to getroot_request_send() in case the getroot()
call will stall.  This is legacy behavior from when flux_msg_copy()
had to be used instead of flux_msg_incref().

Solution: Set the lookup handle as aux data in the message if
we are stalling in lookup_common().  Remove the need to pass the
handle as a parameter to getroot().