resource: conditionally monitor sdmon.online
Problem: nodes are not checked for untracked running work when a
Flux instance starts up.

This might happen, for example, if
- job-exec deems job shell(s) unkillable
- housekeeping/prolog/epilog gets stuck on a hung file system
- the broker exits without proper shutdown

When systemd is enabled, the new sdmon module joins the 'sdmon.online'
broker group on startup.  However, if there are any running flux units,
this is delayed until those units are no longer running.

Change the resource module so that it monitors sdmon.online instead of
broker.online when systemd is enabled.  This will withhold "busy" nodes
from the scheduler until they become idle.

Fixes #6590