broker: update systemd status during shutdown
Problem: when shutdown is held up due due to a small number of
bad nodes, systemctl status doesn't indicate this.

It just says "Waiting for COUNT peers to shutdown" and doesn't
decrement COUNT as the peers disconnect.

Update the systemd message each time the overlay network signals
a change in peer status.