broker: log message when peer won't shut down
Problem: when the instance is stuck in SHUTDOWN state due to
a non-responsive compute node, it's not always evident what is
going on.

Log a message every 60s during shutdown that lists the peer ranks
and hostnames that are still online, e.g.

broker.err[0]: shutdown delayed: waiting for 2 peers: test[42,51] (rank 42,51)

Fixes #6621