broker: trace requests received during shutdown
Problem: when the broker is shutting down, hello requests received
from reconnecting peers are not included in the trace output, but
the resulting disconnect control messages are, leading to confusion.

Trace all requests that cause a disconnect control message to be sent.
Include the (proposed) rank if it is a hello message.

This was discussed in #6621.