resource: add class for handling truncate event
Problem: RFC 44 stipulates that a truncate event shall appear as the
first event in the resource eventlog when it has been truncated, and
that the context of this event shall contain a summary of truncated
information, but the resource eventlog has no support for this event.

Add a truncate class to the resource module which can accumulate
required information from dropped events, providing the base
functionality for a truncate event.