resource: switch to a zlistx_t for resource journal
Problem: The resource journal will need to truncated by dropping
entries from the start, but a jansson json array is not conducive
to this operation.

Swtich to a zlistx_t for the reosurce journal. This data structure
allows the removal and addition of items to the front of the list
(without moving the memory for the rest of the items on the list)