resource: add configurable journal size limit
Problem: The in-memory resource journal grows without bound and there
is no way to limit its size.

In preparation for truncating the resource journal, add a new config
key in the [resource] table, `journal_max`, which can be used to set
a limit on the journal size. Default this limit to 100000 entries.

Modify the reslog constructor to accept this journal_max limit, but
don't do anything with it just yet.