testsuite: improve `flux resource eventlog -F` test
Problem: The test for `flux resource eventlog -F` doesn't check that
the command actually waited for the drain event. There is also a
race between launching the command in the background and draining a
rank where the resource.journal RPC could come after the drain RPC,
thus the command doesn't actually have to follow the eventlog.

Ensure that the RPC is established before moving on to drain a rank
by waiting for output to appear. Check that the drain event is the
final event in the output file.