flux-jobs: filter against all users with --include
Problem: The --include option is predominantly used for finding
jobs on specific hosts.  This implies that the -A option is set (i.e.
filter on all user jobs) but it is currently not.  This can be
confusing to users wondering why --include isn't finding any jobs
running on specific hosts.

Solution: If the user did not specify filtering on a specific user
with --user, then assume all user jobs (i.e. -A or --user=all) will
be checked when --include is specified.

Fixes #6585