python: add option to nullify expansion sentinel in OutputFormat
Problem: The expansion sentinels `+:` and `?+:` allow an output
format field to expand to the maximum width in a given list of
formatted items. However, when a single item is formatted, any minimum
width, as specified in the format spec, will cause the result to be
truncated. This stymies attempts to find identical lines when the
lines are not filtered first.

Add options to the get_format_prepended() and copy() methods to
"nullify" the expansion sentinel by replacing the field without any
width and precision specifiers, i.e. `+:{foo:<3.3}` becomes simply
`{foo}`, which allows the fully rendered width of the field to appear
in any formatted result.