python: fix incorrectly deduplicated lines in Deduplicator
Problem: The Deduplicator class combines lines that would result in
the same result after applying a format string (with some fields
possibly excluded). This doesn't work when the expansion sentinel
`+:` is used, because the Deduplication comes before the expansion,
so the final maximum field width is not known, and instead the minimm
field width is used. This causes unlike lines that differ beyond the
minimum width to be combined.

Pass `nullify_expansion=True` when copying the input format so that the
width is dropped from these fields and deduplication uses the full
width for comparison.

Fixes #6625