flux-version: report hwloc.api=x.y.z instead of hwloc=x.y.z
Problem: The hwloc version reported by `flux version` is the build time
libhwloc API version, not the actual libhwloc version. This can cause
confusion for users that are expecting this string to report the library
version.

Hint that the HWLOC version reported is an API version by changing the
string to `hwloc.api`.

Fixes #6637