resource: shrink down ranks unless method=configuration
Problem: When nodes are lost in a Flux instance the scheduler assumes
the nodes may come back, but this is not supported unless the resources
were defined by configuration (e.g. a system instance). This can
cause jobs submitted to the instance to get stuck in pending state
indefinitely.

Add support for the "shrink" key in the RFC 28 resource acquisition
response. Include lost ranks (those that were online and are now
offline) in the shrink key if the resource-define method is not
"configuration", and remove them from the valid ranks set.