update housekeeping/perilog scripts for transition
Problem: the scripts flux-run-{epilog,prolog,housekeeping} and their
templated systemd unit files are no longer needed.

The job manager will now start the site-provided scripts directly
with the IMP in a transient systemd unit in the Flux systemd instance.

Remove the systemd unit files.

Instead of removing the run scripts and forcing sites to update the IMP
config to run the site scripts immediately, retain the run scripts for now
and modify them to exec the site scripts (making them a no-op).  We can
remove the run scripts later.