housekeeping: let the exec service be configurable
Problem: we'd like to start housekeeping in the flux systemd instance
but it currently hard codes the "rexec" subprocess exec service.

If exec.service is set in configuration, use that for housekeeping too.
If using the IMP with sdexec, set command options for proxy signaling.

Change the bulk exec name from "housekeeping" to "flux-housekeeping".
This becomes the unit name under sdexec, which must begin with "flux-"
to get the required IMP signal forwarding behavior