resource: add resource.monitor-force-down RPC
Problem: When testing with fake resources using the resource module
`monitor-force-up` option, there is no way to simulate lost ranks.

Add a new RPC `resource.monitor-force-down` which takes a payload
of the form `{"ranks":s}`, where `ranks` is an idset of targets to
force down, as if they went from online to offline.

This is only meant for testing, and is meant to be used with the
`monitor-force-up` option.