flux-resource: add `-i, --include=TARGETS` to eventlog command
Problem: It would be useful to filter the resource eventlog by a set
of hosts or broker ranks, but this is not supported, and tools like
`grep` won't work to match a subset of hosts or ranks in an event.

Add a `-i, --include=TARGETS` option to the `flux resource eventlog`
command, where TARGETS is a set of ranks or hosts on which to match.

One caveat is that matching historical drain events by rank may not
give expected results if ranks have been remapped in an instance.