perilog: let the exec service be configurable
Problem: we'd like to start prolog/epilog in the flux systemd instance
but it currently hard codes the "rexec" subprocess exec service.

If exec.service is set in configuration, use that for perilog too.
If using the IMP with sdexec, set command options for proxy signaling.

Change the bulk exec name from prolog/epilog to flux-prolog/flux-epilog".
This becomes the unit name under sdexec, which must begin with "flux-"
to get the required IMP signal forwarding behavior

Although perilog never sends SIGKILL after it was disabled in #6330,
set a kill_signal to SIGUSR1 or SIGKILL in case we want to revert
that change in the future.