broker: extend timeouts for systemctl stop flux
Problem: if a system instance is shut down with 'systemctl stop flux',
as opposed to flux-shutdown(1), the whole stop process is under a 90s
timeout ending in SIGKILL, which sometimes is not enough time to stop a
large system instance cleanly.

To mitigate this, CLEANUP state has a 45s timeout, and in FINALIZE,
flux-dump(1) periodically extends the systemd stop timeout by 10s.
However, if CLEANUP uses up the entire 45s, only 45s is left to split
among the remaining states, more or less.

Experience has shown that 'systemctl stop flux' will be used, so we need
to improve management of the systemd stop timeout, and perhaps even let
it apply to flux-shutdown(1), so the behavior at shutdown is not dependent
on the method used.

As an initial step towards this goal, add a function that effectively
resets the systemd stop timer by calling sd_notify() EXTEND_TIMEOUT_USEC
with a value equal to TimeoutStopSec, which is passed in from the unit
file via broker attribute.  Call this function on entry to each of the
CLEANUP, SHUTDOWN, FINALIZE, and GOODBYE states.  The state transitions
are indicative of progress, and we treat the systemd stop timeout as a
timeout on progress rather than on the entire stop sequence.  This
increases the maximum time allowed to flux for stopping, but ensures the
broker won't be killed while making progress.