sdexec: add a mechanism to time out units
Problem: an sdexec imp-shell unit can run into the following problem:
- flux-shell is killed/terminates
- there are unkillable children of flux-shell
- the IMP won't exit until the cgroup is empty
- the job remains in R state

This adds a configurable stop timer to sdexec that is triggered when
the unit enters deactivating state.  Disabled by default, the timer
is configured via the following subprocess command options:

SDEXEC_STOP_TIMER_SEC
  Specify the timeout value in seconds.  If non-negative, this enables
  the stop timer.
SDEXEC_STOP_TIMER_SIGNAL
  Specify a signal to send to the unit after the timeout.  By default,
  SIGKILL is used.

The behavior of the stop timer is follows:
- The timer is activated when the unit enters "deactivating" state.
- After STOP_TIMER_SEC seconds, STOP_TIMER_SIGNAL is sent to the unit.
- After another STOP_TIMER_SEC seconds, the unit is abandonded and
  subprocess exec RPC is terminated with an EDEADLK error.

To solve the stated problem, the stop timer must be used with job-exec
changes to run the unit with Type=notify, in conjunction with changes
to the IMP to call sd_notify() STOPPING=1 when the shell exits.
The job-exec changes are coming in a future commit.