libsdexec: parse TimeoutStopUSec property
Problem: job-exec may need to set TimeoutStopUSec=infinity to
disable systemd's stop timeout that clobbers the IMP with a SIGKILL.

Since this property has a non-string value, it needs to be special
cased in sdexec_start_transient_unit() so it can be represented using
the proper dbus type.

Add it, with "infinity" support.