perilog: raise non-fatal job exception on epilog failure
Problem: Job epilog failures do not raise job exceptions. This made
sense when the epilog was used for administrative purposes, but now
that most of that has moved to housekeeping, the epilog is meant to
include cleanup necessary for job completion, and therefore an epilog
failure should be reflected in the job eventlog and the user notified.

Raise a non-fatal job exception on epilog failure. A non-fatal
exception is used so that it does influence the job status while
still allowing the exception to be logged.

Fixes #6249