perilog: Add FLUX_JOB_RANKS to environment of prolog/epilog
Problem: The prolog and epilog scripts may need to determine which
node within the job they're running on, but this requires fetching
the job's R or data from the job-list module, which is not a
scalable operation.

Provide the ranks which the job was assigned in a `FLUX_JOB_RANKS`
environment variable to the prolog and epilog. This contains enough
information to derive other attributes of the current node without
fetching data from a central location (e.g. by using the locally
available hostlist attribute)

Fixes #6668