flux-hostlist: add `-F, --find=HOSTS` option
Problem: Other hostlist commands have a `-F, --find` option, but
flux-hostlist(1) does not. Also, there is not another good way to
determine the index of the current host in a hostlist.

Add a `-F, --find=HOSTS` option to flux-hostlist which takes a
hostlist and returns a space-separated list of indices of each
host in the target hostlist.