sdexec: abort module if flux-security < v0.14.0
Problem: The sdexec module launches work as systemd transient
units with Type=notify, which implies flux-security >= 0.14.0 is
being used at runtime. However, we don't want to make this version
of flux-security a hard requirement at this time.

Add a runtime version check to the sdexec module and exit with an
error if an incompatible flux-security version is found.