configure: relax flux-security version check
Problem: configure currently requires flux-security v0.14.0 or later,
but this is only required for the sdexec module, only used in system
instances and a few tests.

Now that the sdexec module and tests have runtime checks for the
flux-security version, relax the required version of flux-security
back to v0.13.0.

Note that v0.14.0 is still installed in CI and required in the debian
control file.