perilog: start timeout timer when launching procesess
Problem: When a prolog/epilog timeout is configured, the perilog.so
plugin delays the start of the timeout timer until the bulk-exec
`on_start` callback is called. But this could result in the timer
never being started if one or more tasks get stuck starting.

Start the timeout timer immediately on launch of the prolog or
epilog tasks. This removes the need for the start_cb(), so remove
that function.

Fixes #6644