resource: ignore torpid/lively events in lost rank detection
Problem: The type of event being processed in post_join_leave() when
adjusting monitor->lost (the set of "lost" ranks) is not checked. This
means that any time `post_join_leave()` is called, the `lost` idset is
adjusted, which is incorrect for all events except "online/offline".
The net effect now is that ranks added to the `torpid` group are
removed from `lost`, and ranks that become `lively` are added to it.
A potentially confusing situation!

Only adjust `monitor->lost` when the `monitor->up` idset is being
modified, i.e. when `oldset == monitor->up`

Fixes #6675