testsuite: ensure torpid/lively events do not affect shrink behavior
Problem: No tests ensures tha torpid/lively events do not affect
detection of lost ranks in batch and alloc jobs.

Add a set of tests to t2317-resource-shrink.t that ensure a node going
from lively to torpid and back again doesn't cause the resources in
a subinstance to shrink.