flux-queue: do not default to --all with configured queues
Problem: The default behavior of `flux queue start|stop` to apply
to all queues when named queues are configured is confusing and
potentially dangerous.

Issue an error instead of a warning when this is the case.

If there are no configured queues, then `flux queue start` and
`flux queue stop` will still work as expected.

Update tests.

Fixes #6674