improve error message on invalid formatting of shell options
Problem: When submitting a job through the command line, a shell
option `-o KEY[=VAL]` is not required to have a value. If one is
not provided, then `attributes.system.shell.KEY` is set to 1. If a
user later provides a sub-key of this shell option, such as `-o
KEY.foo=bar`, the error message returned by Python is not
particularly helpful.

Wrap `set_treedict()` in a try/except where user input might
provide an invalid key and raise a useful error message.
Fixes #6678