broker: add overlay.parent-disconnect RPC
Problem: there is no way to force the broker to disconnect
from its overlay parent, other than from the parent itself.

Add overlay.parent-disconnect RPC, which allows any part of the
system to force an overlay disconnect.  This forces all pending
and future RPCs to the parent to fail fast.  Because the broker
monitors the parent broker state machine using a streaming RPC,
the net effect is to force the subtree to shut down.