sdmon: reconnect when bus is lost
Problem: if streaming RPCs to sdbus fail, for example if the
systemd user instance is temporarily stopped, sdmon treats that
as a fatal error and exits.

The sdbus module was designed to with with EAGAIN when the error
is non-fatal.  If sdbus RPCs fail with EAGAIN, retry the streaming
RPCs for that bus.  They will block until the upstream d-bus connection
is re-established.

This required some internal refactoring:
- instead of one unit hash, use one per bus so it can be easily
  purged when the bus state is invalidated
- separate bus init and bus start functions, so that the latter
  can be called again on restart