flux-queue: do not require reason with `disable`
Problem: `flux queue disable` requires a reason, but this is not
consistent with other `flux queue` subcommands like `flux queue stop`.
It will also make the command less convenient when the reason is
moved to an option argument.

Set a default reason of "disabled by administrator" which is used if
no reason is supplied.

Update a couple tests that supplied nonsense reasons due to the
prior requirement.