flux-queue: allow (most) subcommands to take queue in free args
Problem: Many `flux queue` subcommands take a `-q, --queue=CMD`
option to provide a named queue, but for most of these commands,
providing the queue name as a free argument is the more obvious and
expected behavior.

Allow the status, start, stop, enable and disable subcommands of
flux-queue to take the queue name in the free arguments. Continue to
support the `-q, --queue=NAME` option, but mark it as hidden in the
usage output to allow it to be deprecated in the future.

For stop and disable, move the optional "reason" message to a `-m,
--message=REASON` option, since queue name and reason both can be
specified in free arguments.

Fix usage in testsuite.