doc: update exit-timeout description in flux-shell(1)
Problem: The exit-timeout option is now disabled by default in `flux
alloc` and `flux batch`, but the description of the shell option in
`flux-shell(1)` makes it seem like it is always enabled.

Update the description of exit-timeout in `flux-shell(1)` to indicate
the default for normal parallel jobs differs from jobs that are
instances of Flux.