libsubprocess: remove redundant test in bulk_exec_kill()
Problem: bulk_exec_kill() manually skips subprocesses that are not
in FLUX_SUBPROCESS_RUNNING or FLUX_SUBPROCESS_INIT states, but this
is unnecessary since flux_subprocess_kill() already makes a similar
check and returns ESRCH if the subprocess is not in a valid state for
signaling.  Also, placing this check outside of flux_subprocess_kill()
means that future updates to the determination of when a process can
be signaled within libsubprocess proper won't have an effect. (For
example, a future fix will be forthcoming to allow process groups to
be signaled even if the main process has already exited).

Remove the redundant check for RUNNING/INIT state and ignore inactive
subprocesses (those that return ESRCH from flux_subprocess_kill()).