libsubprocess: allow signal of pgrp after main pid exits
Problem: After the main process exits, libsubprocess no longer allows
signals to be delivered via flux_subprocess_kill(), even if a process
group was created and child processes remain in it. This can result
in an indefinite hang in cases where a libsubprocess client is trying
to kill off a subprocess that created children: The child processes
can't be signaled, but the subprocess is not complete because stdio
is still open.

Update the check for subprocesses eligible for signal delivery to
allow if the subprocess is still active when a process group was
used. If not, keep the previous logic of only allowing running or
initializing subprocesses in flux_subprocess_kill().

Fixes #6712