libsubprocess: add test that signals pgrp after leader exit
Problem: There are no tests in the libsubprocess testsuite that
ensure a process group with remaining processes can be signaled
with flux_subprocess_kill() even after the main pid (the process
group leader) exits.

Add a mode to to test/test_fork_sleep.c that exits in the parent
immediately if TEST_FORK_SLEEP_NOWAIT is set in the environment.
Add a new test simlar to test_kill_setpgrp() that ensures the
caller can still signal the sleeping pid after its parent exits.