shell: clear signal mask and handlers at fork()
Problem: While libsubprocess handles resetting the signal mask and
handlers of tasks spawned by the job shell, this is not done if
a shell plugin calls fork() directly.  This could leave the child
process lingering after the job itself is terminated, since the
process would by default be ignoring SIGTERM.

Set up an atfork handler in the job shell that unblocks all signals
and resets the signal disposition for signals handled by the shell.