python: accept ResourceSet dict in SchedResourceList initializer
Problem: The SchedResourceList initializer assumes it will only be
passed a dictionary of dictionaries, each of which needs to be passed
to the ResourceSet() constructor. This assumption makes it awkward
to create a SchedResourceList from another instance of the same class.

Check to see if the each value is already a ResourceSet in the `resp`
parameter to the SchedResourceList initializer, and only convert to
a ResourceSet if necessary.