python: add QueueList class for fetching queue information
Problem: There's no user friendly method in Python to fetch the list
of configured queues along with information about them, including
name (if named queues are used), status, limits and defaults, and
associated resources.

Add a QueueList class which can fetch information about all queues
using a combination of the configuration, queue statuses, and resource
list. It presents as a list of QueueInfo objects, each of which contains
attributes describing the above data.