flux-queue: support limits.range.effective.{nnodes,ncores,ngpus}
Problem: When there is no limit set for resources, `flux queue list`
reports `inf` as the maximum. However, the real maximum is actually
the number of resources assigned to that queue -- requests for more
than this count will be rejected as unsatisfiable.

Add an "effective" range class which reports the minimum of the
policy limit and all resources in a queue instead of just the policy
limit. Make these effective ranges available as

  limits.range.effective.nnodes
  limits.range.effective.ncores
  limits.range.effective.ngpus